requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
langchain_community
